# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This implements the Airport API Reference.",
        "title": "Airport API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/airports": {
            "get": {
                "description": "Get a list of all airports",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "Get all airports",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all airports",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Airport"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Add a new airport to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "Add a new airport",
                "parameters": [
                    {
                        "description": "Airport details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers_airports.AddAirportRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Airport"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            }
        },
        "/airports/{id}": {
            "get": {
                "description": "Get details of an airport using its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "Get an airport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Airport ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved airport",
                        "schema": {
                            "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Airport"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Update information of an airport given its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "Update airport",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Airport ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body with updated information",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers_airports.UpdateAirportRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated airport information",
                        "schema": {
                            "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Airport"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid or missing parameters",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete an airport from the system using its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "Delete an airport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Airport ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            }
        },
        "/cities": {
            "get": {
                "description": "Get the list of all cities from the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Retrieve all cities",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all cities",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.City"
                            }
                        }
                    },
                    "404": {
                        "description": "No cities found",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Add a new city to the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Add a new city",
                "parameters": [
                    {
                        "description": "City details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers_cities.AddCityBodyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created city object",
                        "schema": {
                            "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.City"
                        }
                    },
                    "400": {
                        "description": "Bad request, JSON decoding failed",
                        "schema": {}
                    },
                    "404": {
                        "description": "Internal server error when saving the city into the database",
                        "schema": {}
                    }
                }
            }
        },
        "/cities/{id}": {
            "get": {
                "description": "Get a specific city from the database by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Retrieve a city by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "City ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the city",
                        "schema": {
                            "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.City"
                        }
                    },
                    "404": {
                        "description": "City not found",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Update the details of a specific city in the database by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Update a city's details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "City ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Name of the city",
                        "name": "name",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "State of the city",
                        "name": "state",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated the city",
                        "schema": {
                            "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.City"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {}
                    },
                    "404": {
                        "description": "City not found",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete a city from the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Delete a city",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "City ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City successfully deleted",
                        "schema": {}
                    },
                    "404": {
                        "description": "City not found",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_krymancer_airport-api-go_pkg_common_models.Airport": {
            "type": "object",
            "properties": {
                "city_id": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "iata_code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_krymancer_airport-api-go_pkg_common_models.City": {
            "type": "object",
            "properties": {
                "airports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Airport"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "handlers_airports.AddAirportRequestBody": {
            "type": "object",
            "properties": {
                "city_id": {
                    "type": "integer",
                    "format": "integer",
                    "example": 1
                },
                "iata_code": {
                    "type": "string",
                    "format": "string",
                    "example": "JFK"
                },
                "name": {
                    "type": "string",
                    "format": "string",
                    "example": "John F. Kennedy International Airport"
                }
            }
        },
        "handlers_airports.UpdateAirportRequestBody": {
            "type": "object",
            "properties": {
                "city_id": {
                    "type": "integer",
                    "format": "integer",
                    "example": 1
                },
                "iata_code": {
                    "type": "string",
                    "format": "string",
                    "example": "JFK"
                },
                "name": {
                    "type": "string",
                    "format": "string",
                    "example": "John F. Kennedy International Airport"
                }
            }
        },
        "handlers_cities.AddCityBodyRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "format": "string",
                    "example": "New York"
                },
                "state": {
                    "type": "string",
                    "format": "string",
                    "example": "NY"
                }
            }
        },
...
    }
}